% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcidata-package.R
\docType{package}
\name{hcidata-package}
\alias{hcidata-package}
\alias{_PACKAGE}
\alias{hcidata}
\title{hcidata: HCI Datasets}
\description{
A collection of datasets from the field of human-computer interaction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/henningpohl/hcidata}
  \item Report bugs at \url{https://github.com/henningpohl/hcidata/issues}
}

}
\author{
\strong{Maintainer}: Henning Pohl \email{HenningPohl@gmail.com} (\href{https://orcid.org/0000-0002-1420-4309}{ORCID})

}
\keyword{internal}
