% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcidata-package.R
\docType{package}
\name{hcidata-package}
\alias{hcidata-package}
\alias{_PACKAGE}
\alias{hcidata}
\title{hcidata: HCI Datasets}
\description{
A collection of datasets of human-computer interaction (HCI) experiments. Each dataset is from an HCI paper, with all fields described and the original publication linked. All paper authors of included data have consented to the inclusion of their data in this package. The datasets include data from a range of HCI studies, such as pointing tasks, user experience ratings, and steering tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/henningpohl/hcidata}
  \item Report bugs at \url{https://github.com/henningpohl/hcidata/issues}
}

}
\author{
\strong{Maintainer}: Henning Pohl \email{HenningPohl@gmail.com} (\href{https://orcid.org/0000-0002-1420-4309}{ORCID})

}
\keyword{internal}
